moved commit_account_changes into journalist_app/utils.py
