+++ securedrop/journalist.py
@@ -22,7 +22,7 @@ import worker
 
 from journalist_app import create_app
 from journalist_app.decorators import login_required, admin_required
-from journalist_app.utils import get_source
+from journalist_app.utils import get_source, commit_account_changes
 
 app = create_app(config)
 
@@ -233,22 +233,6 @@ def admin_reset_two_factor_hotp():
 
 class PasswordMismatchError(Exception):
     pass
-
-
-def commit_account_changes(user):
-    if db_session.is_modified(user):
-        try:
-            db_session.add(user)
-            db_session.commit()
-        except Exception as e:
-            flash(gettext(
-                "An unexpected error occurred! Please check the application "
-                  "logs or inform your adminstrator."), "error")
-            app.logger.error("Account changes for '{}' failed: {}".format(user,
-                                                                          e))
-            db_session.rollback()
-        else:
-            flash(gettext("Account updated."), "success")
 
 
 @app.route('/admin/edit/<int:user_id>', methods=('GET', 'POST'))
